Refactor: Modularize visualize_iteration_history
